The application is installed on the child's device as well as on the parent's device, in the login view the type of user is chosen.
Feacture
Totally free lifetime
Multiple Child clients
Hidden app icon (stealth mode)
Real-time location.
Recording calls: incoming/outgoing.
SMS: received/sent.
Environment recording.
Take pictures.
Keylogger.
Phishing social network.
Notifications received: Whatsapp, Instagram, Messenger.

note: it is very important that accept all the necessary permissions for the application to work properly

Disclaimer

The Is my spy app application is intended for legal and educational purposes ONLY. It is a violation of the law to install surveillance software on a mobile phone that you have no right to monitor.

Is my spy app is not responsible if the user does not follow the laws of the country and goes against it. If it is found that the user violates any law or spy in secret, he will be subject to sanctions that govern the legislation of the country.